refactoring, moved models from api to core
