+++ src/api/v1/product/filters/product.py
@@ -1,6 +1,6 @@
 from django_filters import rest_framework as filters
 
-from ..models import Product
+from core.product.models import Product
 
 
 class ProductFilter(filters.FilterSet):
