+++ src/api/v1/product/serializers/category/request.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from api.v1.product.models import ProductCategory
+from core.product.models import ProductCategory
 
 
 class ProductCategoryCreateRequestSeriliazer(serializers.ModelSerializer):
