+++ src/api/v1/product/serializers/product/response.py
@@ -1,7 +1,7 @@
 from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
 
-from api.v1.product.models import Product
+from core.product.models import Product
 from api.v1.product.serializers.image import ProductImageListResponseSerializer
 from api.v1.product import services
 from .common import ProductOptionListSerializer
