+++ src/core/product/admin.py
