+++ src/core/product/apps.py
@@ -1,7 +1,6 @@
 from django.apps import AppConfig
 
 
-class ProductsConfig(AppConfig):
+class ProductConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "api.v1.product"
-    label = "api_v1_product"
+    name = "core.product"
