+++ src/core/product/constants.py
