+++ src/core/product/models/base.py
