+++ src/core/product/models/image.py
