+++ src/core/product/models/manufacturer.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from .. import constants
+from core.product import constants
 from .base import Base
